get rid of the duplicated file_io
