+++ lib/extras/dec/color_hints.cc
@@ -8,7 +8,7 @@
 #include <jxl/encode.h>
 
 #include "lib/extras/dec/color_description.h"
-#include "lib/jxl/base/file_io.h"
+#include "tools/file_io.h"
 
 namespace jxl {
 namespace extras {
@@ -43,7 +43,7 @@ Status ApplyColorHints(const ColorHints& color_hints,
 
           got_color_space = true;
         } else if (key == "icc_pathname") {
-          JXL_RETURN_IF_ERROR(ReadFile(value, &ppf->icc));
+          JXL_RETURN_IF_ERROR(jpegxl::tools::ReadFile(value, &ppf->icc));
           got_color_space = true;
         } else {
           JXL_WARNING("Ignoring %s hint", key.c_str());
