+++ lib/jxl_lists.bzl
@@ -15,7 +15,6 @@ libjxl_base_sources = [
     "jxl/base/compiler_specific.h",
     "jxl/base/data_parallel.cc",
     "jxl/base/data_parallel.h",
-    "jxl/base/file_io.h",
     "jxl/base/float.h",
     "jxl/base/iaca.h",
     "jxl/base/os_macros.h",
