+++ lib/jxl_lists.cmake
@@ -15,7 +15,6 @@ set(JPEGXL_INTERNAL_BASE_SOURCES
   jxl/base/compiler_specific.h
   jxl/base/data_parallel.cc
   jxl/base/data_parallel.h
-  jxl/base/file_io.h
   jxl/base/float.h
   jxl/base/iaca.h
   jxl/base/os_macros.h
