+++ tools/benchmark/benchmark_utils.cc
@@ -21,8 +21,8 @@
 
 #include <fstream>
 
-#include "lib/jxl/base/file_io.h"
 #include "lib/jxl/image_bundle.h"
+#include "tools/file_io.h"
 
 extern char** environ;
