+++ tools/comparison_viewer/image_loading.cc
@@ -10,10 +10,10 @@
 
 #include "lib/extras/codec.h"
 #include "lib/extras/dec/color_hints.h"
-#include "lib/jxl/base/file_io.h"
 #include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/image_bundle.h"
 #include "lib/jxl/image_metadata.h"
+#include "tools/file_io.h"
 #include "tools/thread_pool_internal.h"
 #include "tools/viewer/load_jxl.h"
