+++ tools/djxl_fuzzer_corpus.cc
@@ -26,7 +26,6 @@
 #include "lib/extras/codec.h"
 #endif
 #include "lib/jxl/base/data_parallel.h"
-#include "lib/jxl/base/file_io.h"
 #include "lib/jxl/base/override.h"
 #include "lib/jxl/base/span.h"
 #include "lib/jxl/codec_in_out.h"
@@ -40,6 +39,7 @@
 #include "lib/jxl/encode_internal.h"
 #include "lib/jxl/jpeg/enc_jpeg_data.h"
 #include "lib/jxl/modular/encoding/context_predict.h"
+#include "tools/file_io.h"
 #include "tools/thread_pool_internal.h"
 
 namespace {
@@ -283,7 +283,7 @@ bool GenerateFile(const char* output_dir, const ImageSpec& spec,
     }
   }
 
-  if (!jxl::WriteFile(compressed, output_fn)) return 1;
+  if (!jpegxl::tools::WriteFile(output_fn, compressed)) return 1;
   if (!quiet) {
     std::unique_lock<std::mutex> lock(stderr_mutex);
     std::cerr << "Stored " << output_fn << " size: " << compressed.size()
